+++ paymentsheet/src/main/java/com/stripe/android/customersheet/ui/CustomerSheetScreen.kt
@@ -14,7 +14,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.stripe.android.common.ui.BottomSheetLoadingIndicator
 import com.stripe.android.common.ui.PrimaryButton
-import com.stripe.android.core.utils.FeatureFlags.customerSheetACHv2
 import com.stripe.android.customersheet.CustomerSheetViewAction
 import com.stripe.android.customersheet.CustomerSheetViewState
 import com.stripe.android.model.PaymentMethodCode
@@ -28,7 +27,6 @@ import com.stripe.android.paymentsheet.ui.PaymentOptions
 import com.stripe.android.paymentsheet.ui.PaymentSheetScaffold
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBar
 import com.stripe.android.paymentsheet.utils.PaymentSheetContentPadding
-import com.stripe.android.ui.core.FormUI
 import com.stripe.android.ui.core.cbc.CardBrandChoiceEligibility
 import com.stripe.android.ui.core.elements.H4Text
 import com.stripe.android.ui.core.elements.SimpleDialogElementUI
@@ -75,18 +73,11 @@ internal fun CustomerSheetScreen(
                         PaymentSheetContentPadding()
                     }
                     is CustomerSheetViewState.AddPaymentMethod -> {
-                        if (customerSheetACHv2.isEnabled) {
-                            AddPaymentMethodWithPaymentElement(
-                                viewState = viewState,
-                                viewActionHandler = viewActionHandler,
-                                displayForm = displayAddForm,
-                            )
-                        } else {
-                            AddPaymentMethod(
-                                viewState = viewState,
-                                viewActionHandler = viewActionHandler,
-                            )
-                        }
+                        AddPaymentMethod(
+                            viewState = viewState,
+                            viewActionHandler = viewActionHandler,
+                            displayForm = displayAddForm,
+                        )
                         PaymentSheetContentPadding()
                     }
                     is CustomerSheetViewState.EditPaymentMethod -> {
@@ -181,50 +172,8 @@ internal fun SelectPaymentMethod(
 }
 
 @Composable
+@Suppress("LongMethod")
 internal fun AddPaymentMethod(
-    viewState: CustomerSheetViewState.AddPaymentMethod,
-    viewActionHandler: (CustomerSheetViewAction) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val horizontalPadding = dimensionResource(R.dimen.stripe_paymentsheet_outer_spacing_horizontal)
-
-    Column(
-        modifier = modifier.padding(horizontal = horizontalPadding),
-    ) {
-        H4Text(
-            text = stringResource(id = R.string.stripe_paymentsheet_save_a_new_payment_method),
-            modifier = Modifier
-                .padding(bottom = 20.dp)
-        )
-
-        FormUI(
-            hiddenIdentifiers = viewState.formViewData.hiddenIdentifiers,
-            enabled = viewState.enabled,
-            elements = viewState.formViewData.elements,
-            lastTextFieldIdentifier = viewState.formViewData.lastTextFieldIdentifier,
-            modifier = Modifier.padding(bottom = 8.dp),
-        )
-
-        AnimatedVisibility(visible = viewState.errorMessage != null) {
-            viewState.errorMessage?.let { error ->
-                ErrorMessage(error = error)
-            }
-        }
-
-        PrimaryButton(
-            label = stringResource(id = R.string.stripe_paymentsheet_save),
-            isEnabled = viewState.primaryButtonEnabled,
-            isLoading = viewState.isProcessing,
-            onButtonClick = {
-                viewActionHandler(CustomerSheetViewAction.OnPrimaryButtonPressed)
-            },
-            modifier = Modifier.padding(top = 10.dp),
-        )
-    }
-}
-
-@Composable
-internal fun AddPaymentMethodWithPaymentElement(
     viewState: CustomerSheetViewState.AddPaymentMethod,
     viewActionHandler: (CustomerSheetViewAction) -> Unit,
     displayForm: Boolean,
@@ -268,7 +217,7 @@ internal fun AddPaymentMethodWithPaymentElement(
                     enabled = viewState.enabled,
                     supportedPaymentMethods = viewState.supportedPaymentMethods,
                     selectedItem = viewState.selectedPaymentMethod,
-                    formElements = viewState.formElements,
+                    formElements = viewState.formViewData.elements,
                     linkSignupMode = null,
                     linkConfigurationCoordinator = null,
                     showCheckboxFlow = flowOf(false),