+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/definitions/AuBecsDebitDefinition.kt
@@ -4,6 +4,7 @@ import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
 import com.stripe.android.lpmfoundations.paymentmethod.AddPaymentMethodRequirement
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
+import com.stripe.android.lpmfoundations.paymentmethod.UiDefinitionFactory
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.elements.SharedDataSpec
@@ -21,15 +22,15 @@ internal object AuBecsDebitDefinition : PaymentMethodDefinition {
 
     override fun requiresMandate(metadata: PaymentMethodMetadata): Boolean = true
 
-    override fun supportedPaymentMethod(
-        sharedDataSpec: SharedDataSpec,
-    ): SupportedPaymentMethod {
-        return SupportedPaymentMethod(
-            paymentMethodDefinition = this,
-            sharedDataSpec = sharedDataSpec,
-            displayNameResource = R.string.stripe_paymentsheet_payment_method_au_becs_debit,
-            iconResource = R.drawable.stripe_ic_paymentsheet_pm_bank,
-            tintIconOnSelection = true,
-        )
-    }
+    override fun uiDefinitionFactory(): UiDefinitionFactory = AuBecsDebitUiDefinitionFactory
+}
+
+private object AuBecsDebitUiDefinitionFactory : UiDefinitionFactory.RequiresSharedDataSpec {
+    override fun createSupportedPaymentMethod(sharedDataSpec: SharedDataSpec) = SupportedPaymentMethod(
+        paymentMethodDefinition = AuBecsDebitDefinition,
+        sharedDataSpec = sharedDataSpec,
+        displayNameResource = R.string.stripe_paymentsheet_payment_method_au_becs_debit,
+        iconResource = R.drawable.stripe_ic_paymentsheet_pm_bank,
+        tintIconOnSelection = true,
+    )
 }