+++ paymentsheet/src/test/java/com/stripe/android/customersheet/CustomerSheetViewModelTest.kt
@@ -46,7 +46,12 @@ import com.stripe.android.paymentsheet.ui.PrimaryButton
 import com.stripe.android.testing.CoroutineTestRule
 import com.stripe.android.testing.FeatureFlagTestRule
 import com.stripe.android.testing.PaymentMethodFactory
+import com.stripe.android.ui.core.elements.CardBillingAddressElement
+import com.stripe.android.ui.core.elements.CardDetailsSectionElement
+import com.stripe.android.ui.core.elements.SaveForFutureUseElement
+import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
+import com.stripe.android.uicore.elements.SectionElement
 import com.stripe.android.uicore.forms.FormFieldEntry
 import com.stripe.android.utils.FakeIntentConfirmationInterceptor
 import kotlinx.coroutines.test.TestCoroutineScheduler
@@ -442,7 +447,7 @@ class CustomerSheetViewModelTest {
     }
 
     @Test
-    fun `When CustomerViewAction#OnAddCardPressed, view state is updated to CustomerViewAction#AddPaymentMethod`() = runTest(testDispatcher) {
+    fun `When CustomerViewAction#OnAddCardPressed, view state is updated to CustomerViewAction#AddPaymentMethod and fields are shwon`() = runTest(testDispatcher) {
         val viewModel = createViewModel(
             workContext = testDispatcher
         )
@@ -451,8 +456,43 @@ class CustomerSheetViewModelTest {
             assertThat(awaitItem())
                 .isInstanceOf(SelectPaymentMethod::class.java)
             viewModel.handleViewAction(CustomerSheetViewAction.OnAddCardPressed)
+
+            val item = awaitItem()
+            assertThat(item).isInstanceOf(AddPaymentMethod::class.java)
+
+            val formElements = item.asAddState().formViewData.elements
+
+            assertThat(formElements[0]).isInstanceOf(CardDetailsSectionElement::class.java)
+            assertThat(formElements[1]).isInstanceOf(SectionElement::class.java)
+            assertThat(formElements[1].asSectionElement().fields[0])
+                .isInstanceOf(CardBillingAddressElement::class.java)
+            assertThat(formElements[2]).isInstanceOf(SaveForFutureUseElement::class.java)
+        }
+    }
+
+    @Test
+    fun `When CustomerViewAction#OnAddCardPressed & ACHv2 disabled, view state is updated to CustomerViewAction#AddPaymentMethod and fields are shwon`() = runTest(testDispatcher) {
+        featureFlagTestRule.setEnabled(false)
+
+        val viewModel = createViewModel(
+            workContext = testDispatcher
+        )
+
+        viewModel.viewState.test {
             assertThat(awaitItem())
-                .isInstanceOf(AddPaymentMethod::class.java)
+                .isInstanceOf(SelectPaymentMethod::class.java)
+            viewModel.handleViewAction(CustomerSheetViewAction.OnAddCardPressed)
+
+            val item = awaitItem()
+            assertThat(item).isInstanceOf(AddPaymentMethod::class.java)
+
+            val formElements = item.asAddState().formViewData.elements
+
+            assertThat(formElements[0]).isInstanceOf(CardDetailsSectionElement::class.java)
+            assertThat(formElements[1]).isInstanceOf(SectionElement::class.java)
+            assertThat(formElements[1].asSectionElement().fields[0])
+                .isInstanceOf(CardBillingAddressElement::class.java)
+            assertThat(formElements[2]).isInstanceOf(SaveForFutureUseElement::class.java)
         }
     }
 
@@ -2899,6 +2939,14 @@ class CustomerSheetViewModelTest {
         return CollectBankAccountResultInternal.Completed(
             response = bankAccountResponse,
         )
+    }
+
+    private fun CustomerSheetViewState.asAddState(): AddPaymentMethod {
+        return this as AddPaymentMethod
+    }
+
+    private fun FormElement.asSectionElement(): SectionElement {
+        return this as SectionElement
     }
 
     @Suppress("UNCHECKED_CAST")