+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## XX.XX.XX - 2023-XX-XX
 
+## 20.40.1 - 2024-03-21
+
 ### PaymentSheet
-* [Added] Support for Klarna with SetupIntents and PaymentIntents with `setup_future_usage`.
+* [Added][8132](https://github.com/stripe/stripe-android/pull/8132) Support for Klarna with SetupIntents and PaymentIntents with `setup_future_usage`.
+* [Fixed][8142](https://github.com/stripe/stripe-android/pull/8142) Fixed an issue with `CustomerSheet` where the form fields were not displayed when adding a payment method.
 
 ## 20.40.0 - 2024-03-18